pyproject.toml
[tool.poetry]
name = "engage-engine.static"
version = "1.0.0"
description = ""
authors = ["BD103 <dont@stalk.me>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.8"
jinja2 = "^2.11.3"
libsass = "^0.20.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
